pyproject.toml
[project]
name = "integration-job-searcher"
version = "0.1.0"
description = "Job search and processing flow with dbt"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "prefect>=3.4.24",
    "pandas",
    "requests",
    "python-jobspy",
    "prefect-dbt>=0.7.0",
    "sqlalchemy",
    "anthropic",
    "beautifulsoup4",
    "lxml",
    "psycopg2-binary==2.9.9",
    "dbt-postgres>=1.8.1",
    "rapidfuzz",
    "llm-queue @ file:///${PROJECT_ROOT}/../llm-queue/client",
]

[tool.uv]
override-dependencies = [
    "psycopg2-binary==2.9.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
